Update GOagentJS.php
"*Issue 12 - *To show name and number after ending a call using manual
hangup. This is because our users tend to forget copying the name and
number before ending the call for an Appointment set. Once they end the
call using Hotkey, the lead details will disappear. When they use manual
hang up, at least the name and number will still be available for them to
copy. Is there a way that the blind can also copy the name and number from
here?"